Update charter based on F2F discussions during TPAC
The following changes were made:
- Completion dates were updated to reflect current implementation plans, and
the dependency on Open Screen Protocol was noted in particular to justify the
completion date of the Presentation API in Q4 2019.
- The intent to work on a testing API was noted in Test suite section. The text
gives freedom to the Working Group to put that testing API in the current drafts
or in Level 2 specs.
- The reference to the Network Service Discovery API was dropped
- The "boilerplate" text was updated to match latest template, in particular
in the coordination section (I did not try to inject all template updates
though, so as to keep diff minimal)
- The liaisons section was adjusted.