Put Open Screen Protocol in scope
This update puts the Open Screen Protocol in scope of the Working Group:
- The scope section was updated to mention the protocol, describe its scope,
and the willingness to reuse existing network protocols for discovery,
authentication, transport and messaging. It leaves the door open to moving
parts of the protocol to IETF if needed. It maintains the "informative" link
between the APIs and the protocol.
- The out of scope section was adjusted accordingly.
- The Open Screen Protocol is now mentioned as a deliverable of the group
- The liaison with IETF was completed to explicitly mention IETF groups that
develop or maintain protocols used in the Open Screen Protocol.
- Other references to the Open Screen Protocol and the Community Group updated
as needed.