Misc. updates from internal review (#17)
* Update team effort, add ref drafts, drop WPWG

- Adjust team contact effort to better match reality
- Add reference drafts for deliverables per process
- Drop mentions of Web Platform Working Group, which no longer exists.

* Add explicit reference to fingerprinting surface

* Add milestones for Open Screen Protocol

* Clarify that UA need to know about application-specific media

The sentence about application-specific media seems to suggest that the
controlling page can initiate rendering of application-specific media that is
totally opaque to the user agent. In practice, the user agents need to support
(and agree with sending/receiving) the media that is going to be presented.

* Clarify that technical work is primarily done on GitHub

* Clarify in goals that OSP is an application of existing IETF protocols

Goal is to be clear right from the start that the suite of network protocols
that the group will define is an application of existing IETF protocols
(note this already appears further down in the scope section).