Clarify scope of control protocol in OSP (#44)
Via https://github.com/w3c/strategy/issues/444#issuecomment-2017745456

Goal is to clarify that the control protocol targets messages requiresd to
implement the APIs, and not e.g., all of Matter. Note the Out of scope section
is already clear that protocols "not required to implement the APIs developed
by this Working Group" are out of scope.